moving server files into server folder (except for main server.js)
